feature: 채팅방 API 업데이트
- 채팅방 목록 가져올 때 최근 메시지 및 안 읽은 메시지 횟수 추가

+++ backend/ws/ws.go
@@ -20,7 +20,7 @@ type WSTestRequest struct {
 type WSChatEvent struct {
 	ChatRoomID  int
 	ChatMsg     models.ChatMsg
-	UnreadCount int
+	UnreadCount int64
 }
 
 type WSNotiEvent struct {
